pyproject.toml
[project]
name = "adk-google-docs-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Moti Elmakyes",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "google-adk (>=1.13.0,<2.0.0)",
    "rich>=13.0.0,<14.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pydantic>=2.0.0,<3.0.0"
]

[tool.poetry]
packages = [
    { include = "src" }
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
